used gist exmple of Popup + Progress Bar to show loading progress
Loading the autofinger list seems to be the only task (at the
moment) that is slower to do than loading the popup.  As such, I
removed other mentions of pop-up loading windows because the flash
that you can never read may be too distracting/confusing.  This
gist can always be used an example to extend loading to other
API calls (probably especially posting an edit) in the future as
needed.