refactored api calls;added version to login; log to disk
I refactored most of the API calls.  First, I pulled out the
direct mentions of the API urls and moved those to a template
dictionary.  That will make it easier to change the root domain
for testing eventually.  It should also make references to the
particular URL more semantic and easier for (humans) to parse.

Part of refactoring included creating a new function for
get_json_list to handle all the calls around pulling and
updating the autofinger list.

Also the login calls are in a separate function now.

Both of these have vestiges of some problems that crept
in because neither function is incredibly robust.

LLOOGG messages now are also written to disk locally.

Move where the __version__ is defined to make it super easy to
post the version to the login screen.  This will need to be
moved again when we add an about screen.

Added some bits of robustness.  For instance, check if there is
a session file on disk before clearing it out.