create labels on the fly for every chunk of a plan
Because labels have a max content size, we need to split a
plan content into chunks that can be displayed each in a
separate label, automatically added to the gridlayout
containing the plan.  This fix introduces some oddities
with certain types of mark-up (like ref, but I don't
know how ref would show up in a plan anyway) to be dealt
with later. I also broke the planlove clickability.