scroll back to top; bouncback failed searches; better demo scripts
When moving to a new plan, scroll the screen back to the top
of the page.

When we fail a search for a user, it's because we have been given
an invalid username. Track this and bounce the app back to the
search screen with an appropriate error message in the new
(normally empty) label. All other avenues for hitting the readTask
function also have the source tracked, just in case we want to
use it for something fancy later.

We added more unique entries for each of the demo available
user accounts.